+++ railties/CHANGELOG.md
@@ -1,10 +1,18 @@
-*   Add `Application#message_verifier` method to return a application's message verifier.
+*   Add `Application#message_verifier` method to return a message verifier.
 
     This verifier can be used to generate and verify signed messages in the application.
 
         message = Rails.application.message_verifier.generate('my sensible data')
         Rails.application.message_verifier.verify(message)
         # => 'my sensible data'
+
+    It is recommended to not use the same verifier to different things, so you can get different
+    verifiers passing the name argument.
+
+        message = Rails.application.message_verifier('cookies').generate('my sensible cookie data')
+
+    By default all the verifiers will share the same salt, so messages generates by one can be
+    verifier by another one.
 
     See the `ActiveSupport::MessageVerifier` documentation to more information.